refactor(errors): descriptive template names
