+++ tests/pytest/core/test_middleware_healthcheck_user_agent.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 import pytest
 
 from benefits.core import middleware
-from benefits.core.views import ROUTE_INDEX, TEMPLATE_INDEX
+from benefits.core.views import TEMPLATE_INDEX
 
 
 @pytest.mark.django_db
@@ -13,7 +13,7 @@ def test_healthcheck_user_agent(mocker, user_agent):
     mocker.patch.object(middleware.settings, "HEALTHCHECK_USER_AGENTS", [user_agent])
     client = Client(HTTP_USER_AGENT=user_agent)
 
-    response = client.get(reverse(ROUTE_INDEX))
+    response = client.get(reverse(middleware.ROUTE_INDEX))
 
     assert response.status_code == 200
     assert response.content.decode("UTF-8") == "Healthy"
@@ -23,7 +23,7 @@ def test_healthcheck_user_agent(mocker, user_agent):
 def test_non_healthcheck_user_agent(mocker, client):
     mocker.patch.object(middleware.settings, "HEALTHCHECK_USER_AGENTS", ["AlwaysOn"])
 
-    response = client.get(reverse(ROUTE_INDEX))
+    response = client.get(reverse(middleware.ROUTE_INDEX))
 
     assert response.status_code == 200
     assert response.template_name == TEMPLATE_INDEX
