+++ tests/pytest/core/test_session.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 import pytest
 
 from benefits.core import models, session
-from benefits.core.views import ROUTE_INDEX
+from benefits.core.middleware import ROUTE_INDEX
 
 
 @pytest.mark.django_db
